backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymongo==4.7.1
scikit-learn==1.3.1
numpy==1.24.4
PyPDF2==3.0.1
python-dotenv==1.0.0
